Move all logic to the async part
