+++ app/CommandHandling/CommandBusServiceProvider.php
@@ -14,7 +14,7 @@ use CultuurNet\UDB3\Event\EventCommandHandler;
 use CultuurNet\UDB3\Event\Productions\ProductionCommandHandler;
 use CultuurNet\UDB3\Labels\LabelServiceProvider;
 use CultuurNet\UDB3\Log\SocketIOEmitterHandler;
-use CultuurNet\UDB3\Mailer\Ownership\SentOwnershipMailHandler;
+use CultuurNet\UDB3\Mailer\Ownership\SentOwnershipMailCommandHandler;
 use CultuurNet\UDB3\Ownership\CommandHandlers\ApproveOwnershipHandler;
 use CultuurNet\UDB3\Ownership\CommandHandlers\DeleteOwnershipHandler;
 use CultuurNet\UDB3\Ownership\CommandHandlers\RejectOwnershipHandler;
@@ -268,7 +268,7 @@ final class CommandBusServiceProvider extends AbstractServiceProvider
             'mails_command_bus_out',
             function () {
                 $commandBus = $this->createResqueCommandBus('mails', $this->container);
-                $commandBus->subscribe($this->container->get(SentOwnershipMailHandler::class));
+                $commandBus->subscribe($this->container->get(SentOwnershipMailCommandHandler::class));
                 return $commandBus;
             }
         );
